refactor(#112): PlanInfoCard.tsx 타입 파일 분리/ 생성
